feat: add keyboard navigation option
Allows users to navigate the grid using their iPad keyboard (if attached). This will update a value in the metadata object allowing for styling on the tile. This is an opt-in feature.

Work in progress.

Known issue when navigating around whitespace
Known issue with buttons being visible (will be hidden)
Known issue with arrow keys not being usable